cleans out comments, cleans readme, cleans unnecessary code
