+++ README.md
@@ -1,26 +1,62 @@
 # Qlean-Quotes
-house cleaning application
-Overview
-Our group will be working on a project about housekeeping. The project's name is Qlean Quotes. Our app is going to provide an user login that user will be able to schedule, purchase the packages with discounts, select the providers and pay for the package directly from the app. When users signed up, they will enter the required info and see the estimate. Once paid, user can view all orders/paid invoices. Also, user will be able to leave a feedback.
+### An app to make housekeeping breezier
 
-Features
-User login : Users will be able to create accounts and log in to the application.Each user will have a profile page. Payment option / Secure payment : Personal information is protected during payment processing. Calendar/Instant Booking : User will be able to see availability on the calendar to book. Calculates Quote : Calculating of the price to provide to user based on given information (square foot, rooms, level of cleanliness). Feedback : User will be able to create a comment/feedback.
-
-Back End Technologies
-Java, Spring boot, Hibernate, mySQL
-
-Front End Technologies
-HTML CSS Javascript BootStrap SASS
-
-What I'll Have to Learn
-Spring Security, Square API, And we will need to learn how to connect front-end and back-end
-
-Project Tracker
-https://trello.com/b/69tSr3fF/qleanquotes
+## Overview
+Our group will be working on a project about housekeeping. 
+The project's name is Qlean Quotes. 
+Our app is going to provide a user login that is secured through Spring Security,
+a user profile and quote history,
+a form that calculates what it would cost to clean their houses, 
+and a Square integration to allow payment and invoicing.
 
 
-Spring Security Learnin' Trail
-Tutorial 1: https://www.youtube.com/watch?v=bB6A490Uh5M (unsuccessful)
-Tutorial 2: https://www.youtube.com/watch?v=E6P7V5vvkus ()
-Tutorial 3, most helpful tutorial so far, has code to create an admin user as well, but also uses deprecated code in securityconfig: https://studygyaan.com/spring-boot/login-register-example-using-spring-boot
-Tutorial 4: also helpful: https://www.tutorialspoint.com/spring_security/spring_security_form_login_with_database.htm
+## Features
+* User login
+* User registration
+* A user profile 
+* Quote calculation form 
+* Payment option / Secure payment : Personal information is protected during payment processing. 
+
+
+## Back End Technologies
+* Java 
+* Spring boot 
+* Spring Security
+* Square
+* Hibernate
+* mySQL
+
+## Front End Technologies
+* HTML 
+* CSS 
+* Javascript 
+* BootStrap 
+* SASS
+
+
+### Project Tracker
+[Group Trello](https://trello.com/b/69tSr3fF/qleanquotes)
+
+
+### Group Drive
+[Group Drive](https://drive.google.com/drive/folders/1nWK7iq8kKH5hRbClgYeDFaXrIuvNv76m)
+
+
+### Wireframes
+[uizard](https://app.uizard.io/p/a0efda59)
+
+
+### Initial MockUps
+[Esra's MockUps](https://drive.google.com/drive/folders/1BG0iARKKRWU9PiPTA0L1JxVI6dpbsX9c)
+
+
+### Spring Security Learning Documentation
+1. [Tutorial 1](https://www.youtube.com/watch?v=bB6A490Uh5M)  : This video was not successful, but it did provide a foundation
+2. [Tutorial 2](https://www.youtube.com/watch?v=E6P7V5vvkus)  : This video was not successful, but it did provide a foundation
+3. [Tutorial 3](https://studygyaan.com/spring-boot/login-register-example-using-spring-boot) : This was the most helpful tutorial and was the first one that configured an authentication provider DAO. This one has code to create an admin user as well, which could be useful in the future. Features deprecated code though. 
+4. [Tutorial 4](https://www.tutorialspoint.com/spring_security/spring_security_form_login_with_database.htm) : This was also helpful, but also uses deprecated code in securityconfig. 
+5. Beyond that, hitting cmd+k on classes and fn+f1 on methods and variables opens the actual java documentation and helps get the most recent info and demo configurations. This helps find demo code to get around the deprecated code examples in the tutorials. 
+6. Add `logging.level.org.springframework.security=DEBUG` to application.properties to get a more verbose console output while configuring security, it will show errors better than without
+
+
+### What's next?
