+++ src/main/resources/templates/authentication/login.html
@@ -6,7 +6,7 @@
 <nav th:replace="~{fragments :: page-header}"></nav>
 
 <h1>Log In</h1>
-
+<!--if logout, the below th:if condition is triggered, find this in securityconfig, logoutURL() query parameter-->
 <div th:if="${param.logout}"> You have logged out. </div>
 
 <form id="login-form" method="post" th:action="@{/authentication/login/process}">
@@ -24,13 +24,10 @@
             <input class="form-control" name="password" type="password"/>
         </label>
     </div>
-
     <div class="form-group">
     <input type="submit" class="btn btn-primary" value="Log In" />
-
      <a href="/authentication/register" class="btn btn-success">Sign Up</a>
     </div>
-
 </form>
 
 <div th:replace="~{fragments :: page-about}"></div>
@@ -40,27 +37,3 @@
 </body>
 </html>
 
-
-
-<!--<form id="login-form" method="post">-->
-<!--    <div class="form-group">-->
-<!--        <label>Email-->
-<!--            <input class="form-control" th:field="${loginFormDTO.email}" />-->
-<!--        </label>-->
-<!--        <p class="text-danger" th:errors="${loginFormDTO.email}"></p>-->
-<!--    </div>-->
-<!--    <div class="form-group">-->
-<!--        <label>Password-->
-<!--            <input class="form-control" th:field="${loginFormDTO.password}" type="password" />-->
-<!--        </label>-->
-<!--        <p class="text-danger" th:errors="${loginFormDTO.password}"></p>-->
-<!--    </div>-->
-
-<!--    <div class="form-group">-->
-<!--        <input type="submit" class="btn btn-primary" value="Log In" />-->
-
-<!--        <a href="/register" class="btn btn-success">Sign Up</a>-->
-<!--    </div>-->
-
-
-<!--</form>-->
