+++ src/main/resources/templates/createquotes.html
@@ -27,8 +27,6 @@
         </label>
         <p class="text-danger" th:errors="${createQuoteFormDTO.numOfBathroom}"></p>
     </div>
-
-
     <div class="form-group">
         <button type="submit" name="calculate" class="btn btn-primary">Calculate</button>
         <a href="/payment" class="btn btn-success">Pay</a>
@@ -37,8 +35,6 @@
         Total cost:$ <strong th:text="${totalCost}"></strong>
     </div>
 </form>
-
-
 </body>
 </html>
